requirements.txt
# ----------------------------
# Core NLP / Retrieval (Phase 2+)
# ----------------------------
transformers>=4.36,<5.0
langchain>=0.1,<0.3
chromadb>=0.4,<0.5

# ----------------------------
# OCR (Primary Engine - GPU)
# ----------------------------
paddlepaddle-gpu==2.6.1.post118
paddleocr>=2.7.0


# ----------------------------
# PDF Processing
# PaddleOCR requires PyMuPDF < 1.21
# ----------------------------
pymupdf>=1.19,<1.21
pdf2image>=1.16,<2.0

# ----------------------------
# Computer Vision (Docker-safe)
# ----------------------------
opencv-python-headless>=4.10,<4.12

# ----------------------------
# Utilities
# ----------------------------
python-dotenv>=1.0,<2.0
numpy>=1.26,<1.27
Pillow>=9.5,<11.0

# ----------------------------
# Streamlit dependency
# ----------------------------
streamlit
pymupdf>=1.19,<1.21
